Implemented regressions described in issues #2,5
Replaced dependance on an id mask in each component type with
an implicit indexing scheme based on their order in the
parameter pack passed to EntityManager.
Replaced std::vector with std::array for the indicies member
of the entity class.
Implementing these regressions fixed several segfault causing issues,
and the code is a little easier to read now as well.